+++ src/Saunter/Options/Filters/IChannelFilter.cs
@@ -1,0 +1,6 @@
+﻿using LEGO.AsyncAPI.Models;
+
+public interface IChannelFilter
+{
+    void Apply(AsyncApiChannel channel, ChannelFilterContext context);
+}