+++ src/Saunter/Options/Filters/IDocumentFilter.cs
@@ -1,0 +1,6 @@
+﻿using LEGO.AsyncAPI.Models;
+
+public interface IDocumentFilter
+{
+    void Apply(AsyncApiDocument document, DocumentFilterContext context);
+}