+++ src/Saunter/Options/Filters/IOperationFilter.cs
@@ -1,0 +1,6 @@
+﻿using LEGO.AsyncAPI.Models;
+
+public interface IOperationFilter
+{
+    void Apply(AsyncApiOperation operation, OperationFilterContext context);
+}