+++ src/Saunter/SharedKernel/Interfaces/IAsyncApiSchemaGenerator.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using LEGO.AsyncAPI.Models;
+
+namespace Saunter.SharedKernel.Interfaces
+{
+    public interface IAsyncApiSchemaGenerator
+    {
+        AsyncApiSchema? Generate(Type? type);
+    }
+}