+++ test/Saunter.Tests/Generation/DocumentGeneratorTests/ClassAttributesTests.cs
@@ -2,7 +2,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Reflection;
-using Saunter.AsyncApiSchema.v2;
 using Saunter.Attributes;
 using Saunter.Generation;
 using Shouldly;