+++ test/Saunter.Tests/Generation/DocumentGeneratorTests/MethodAttributesTests.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Linq;
 using System.Reflection;
-using Microsoft.Extensions.Options;
-using Saunter.AsyncApiSchema.v2;
 using Saunter.Attributes;
 using Saunter.Generation;
 using Shouldly;