+++ test/Saunter.Tests/Generation/Filters/DocumentFilterTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Reflection;
-using Saunter.AsyncApiSchema.v2;
 using Saunter.Generation;
 using Saunter.Generation.Filters;
 using Shouldly;
@@ -49,12 +48,12 @@ namespace Saunter.Tests.Generation.Filters
         {
             public void Apply(AsyncApiDocument document, DocumentFilterContext context)
             {
-                var channel = new ChannelItem
+                var channel = new AsyncApiChannel
                 {
                     Description = "an example channel for testing"
                 };
 
-                document.Channels.Add(new KeyValuePair<string, ChannelItem>("foo", channel));
+                document.Channels.Add(new KeyValuePair<string, AsyncApiChannel>("foo", channel));
             }
         }
     }